requirements.txt
# Airflow Core
apache-airflow==2.8.0
apache-airflow-providers-apache-spark==4.1.3

# Data Processing
pandas==2.0.3
numpy==1.24.3
delta-spark==3.0.0
pyspark==3.5.0

# S3/MinIO Support
boto3==1.28.85
minio==7.1.17

# Database
psycopg2-binary==2.9.9
pymysql==1.1.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
